+++ CHANGELOG.md
@@ -78,3 +78,9 @@ All notable changes to the "php-namespace-resolver" extension will be documented
 ## 0.4.2
 
 - fix not parsing of other types than class
+
+## 0.4.3
+
+- fix giving error when opening invalid workspace
+- fix not generating namespace
+- add new option `namespaceResolver.forceReplaceSimilarImports` to replace similar class import instead of keeping both (old & new)